+++ examples/ime.rs
@@ -6,7 +6,7 @@ use winit::{
     dpi::PhysicalPosition,
     event::{ElementState, Event, Ime, VirtualKeyCode, WindowEvent},
     event_loop::{ControlFlow, EventLoop},
-    window::WindowBuilder,
+    window::{ImePurpose, WindowBuilder},
 };
 
 fn main() {
@@ -18,6 +18,7 @@ fn main() {
     println!("IME position will system default");
     println!("Click to set IME position to cursor's");
     println!("Press F2 to toggle IME. See the documentation of `set_ime_allowed` for more info");
+    println!("Press F3 to cycle through IME purposes.");
 
     let event_loop = EventLoop::new();
 
@@ -26,6 +27,7 @@ fn main() {
         .build(&event_loop)
         .unwrap();
 
+    let mut ime_purpose = ImePurpose::Normal;
     let mut ime_allowed = true;
     window.set_ime_allowed(ime_allowed);
 
@@ -67,7 +69,7 @@ fn main() {
                 event: WindowEvent::Ime(event),
                 ..
             } => {
-                println!("{:?}", event);
+                println!("{event:?}");
                 may_show_ime = event != Ime::Disabled;
                 if may_show_ime {
                     window.set_ime_position(ime_pos);
@@ -77,20 +79,31 @@ fn main() {
                 event: WindowEvent::ReceivedCharacter(ch),
                 ..
             } => {
-                println!("ch: {:?}", ch);
+                println!("ch: {ch:?}");
             }
             Event::WindowEvent {
                 event: WindowEvent::KeyboardInput { input, .. },
                 ..
             } => {
-                println!("key: {:?}", input);
+                println!("key: {input:?}");
 
                 if input.state == ElementState::Pressed
                     && input.virtual_keycode == Some(VirtualKeyCode::F2)
                 {
                     ime_allowed = !ime_allowed;
                     window.set_ime_allowed(ime_allowed);
-                    println!("\nIME: {}\n", ime_allowed);
+                    println!("\nIME allowed: {ime_allowed}\n");
+                }
+                if input.state == ElementState::Pressed
+                    && input.virtual_keycode == Some(VirtualKeyCode::F3)
+                {
+                    ime_purpose = match ime_purpose {
+                        ImePurpose::Normal => ImePurpose::Password,
+                        ImePurpose::Password => ImePurpose::Terminal,
+                        _ => ImePurpose::Normal,
+                    };
+                    window.set_ime_purpose(ime_purpose);
+                    println!("\nIME purpose: {ime_purpose:?}\n");
                 }
             }
             _ => (),