+++ examples/request_redraw_threaded.rs
@@ -25,7 +25,7 @@ fn main() {
     });
 
     event_loop.run(move |event, _, control_flow| {
-        println!("{:?}", event);
+        println!("{event:?}");
 
         control_flow.set_wait();