+++ src/platform_impl/android/mod.rs
@@ -25,7 +25,9 @@ use crate::{
     error,
     event::{self, StartCause, VirtualKeyCode},
     event_loop::{self, ControlFlow, EventLoopWindowTarget as RootELW},
-    window::{self, CursorGrabMode, ResizeDirection, Theme, WindowButtons, WindowLevel},
+    window::{
+        self, CursorGrabMode, ImePurpose, ResizeDirection, Theme, WindowButtons, WindowLevel,
+    },
 };
 
 static HAS_FOCUS: Lazy<RwLock<bool>> = Lazy::new(|| RwLock::new(true));
@@ -349,10 +351,10 @@ impl<T: 'static> EventLoop<T> {
         }
     }
 
-    fn single_iteration<'a, F>(
+    fn single_iteration<F>(
         &mut self,
         control_flow: &mut ControlFlow,
-        main_event: Option<MainEvent<'a>>,
+        main_event: Option<MainEvent<'_>>,
         pending_redraw: &mut bool,
         cause: &mut StartCause,
         callback: &mut F,
@@ -1005,6 +1007,8 @@ impl Window {
     pub fn set_ime_position(&self, _position: Position) {}
 
     pub fn set_ime_allowed(&self, _allowed: bool) {}
+
+    pub fn set_ime_purpose(&self, _purpose: ImePurpose) {}
 
     pub fn focus_window(&self) {}