+++ src/platform_impl/macos/window.rs
@@ -19,8 +19,9 @@ use crate::{
         LogicalPosition, LogicalSize, PhysicalPosition, PhysicalSize, Position, Size, Size::Logical,
     },
     error::{ExternalError, NotSupportedError, OsError as RootOsError},
+    event::WindowEvent,
     icon::Icon,
-    platform::macos::WindowExtMacOS,
+    platform::macos::{OptionAsAlt, WindowExtMacOS},
     platform_impl::platform::{
         app_state::AppState,
         appkit::NSWindowOrderingMode,
@@ -32,8 +33,8 @@ use crate::{
         Fullscreen, OsError,
     },
     window::{
-        CursorGrabMode, CursorIcon, ResizeDirection, Theme, UserAttentionType, WindowAttributes,
-        WindowButtons, WindowId as RootWindowId, WindowLevel,
+        CursorGrabMode, CursorIcon, ImePurpose, ResizeDirection, Theme, UserAttentionType,
+        WindowAttributes, WindowButtons, WindowId as RootWindowId, WindowLevel,
     },
 };
 use core_graphics::display::{CGDisplay, CGPoint};
@@ -84,6 +85,7 @@ pub struct PlatformSpecificWindowBuilderAttributes {
     pub disallow_hidpi: bool,
     pub has_shadow: bool,
     pub accepts_first_mouse: bool,
+    pub option_as_alt: OptionAsAlt,
 }
 
 impl Default for PlatformSpecificWindowBuilderAttributes {
@@ -99,6 +101,7 @@ impl Default for PlatformSpecificWindowBuilderAttributes {
             disallow_hidpi: false,
             has_shadow: true,
             accepts_first_mouse: true,
+            option_as_alt: Default::default(),
         }
     }
 }
@@ -156,6 +159,9 @@ pub struct SharedState {
     /// transitioning back to borderless fullscreen.
     save_presentation_opts: Option<NSApplicationPresentationOptions>,
     pub current_theme: Option<Theme>,
+
+    /// The state of the `Option` as `Alt`.
+    pub(crate) option_as_alt: OptionAsAlt,
 }
 
 impl SharedState {
@@ -216,7 +222,7 @@ impl WinitWindow {
         }
 
         let this = autoreleasepool(|_| {
-            let screen = match &attrs.fullscreen {
+            let screen = match attrs.fullscreen.clone().map(Into::into) {
                 Some(Fullscreen::Borderless(Some(monitor)))
                 | Some(Fullscreen::Exclusive(VideoMode { monitor, .. })) => {
                     monitor.ns_screen().or_else(NSScreen::main)
@@ -367,6 +373,8 @@ impl WinitWindow {
                     this.center();
                 }
 
+                this.set_option_as_alt(pl_attrs.option_as_alt);
+
                 Id::into_shared(this)
             })
         })
@@ -458,19 +466,25 @@ impl WinitWindow {
         let delegate = WinitWindowDelegate::new(&this, attrs.fullscreen.is_some());
 
         // Set fullscreen mode after we setup everything
-        this.set_fullscreen(attrs.fullscreen);
+        this.set_fullscreen(attrs.fullscreen.map(Into::into));
 
         // Setting the window as key has to happen *after* we set the fullscreen
         // state, since otherwise we'll briefly see the window at normal size
         // before it transitions.
         if attrs.visible {
-            // Tightly linked with `app_state::window_activation_hack`
-            this.makeKeyAndOrderFront(None);
+            if attrs.active {
+                // Tightly linked with `app_state::window_activation_hack`
+                this.makeKeyAndOrderFront(None);
+            } else {
+                this.orderFront(None);
+            }
         }
 
         if attrs.maximized {
             this.set_maximized(attrs.maximized);
         }
+
+        delegate.queue_event(WindowEvent::Focused(false));
 
         Ok((this, delegate))
     }
@@ -1148,6 +1162,9 @@ impl WinitWindow {
     }
 
     #[inline]
+    pub fn set_ime_purpose(&self, _purpose: ImePurpose) {}
+
+    #[inline]
     pub fn focus_window(&self) {
         let is_minimized = self.isMiniaturized();
         let is_visible = self.isVisible();
@@ -1338,6 +1355,16 @@ impl WindowExtMacOS for WinitWindow {
 
     fn set_document_edited(&self, edited: bool) {
         self.setDocumentEdited(edited)
+    }
+
+    fn set_option_as_alt(&self, option_as_alt: OptionAsAlt) {
+        let mut shared_state_lock = self.shared_state.lock().unwrap();
+        shared_state_lock.option_as_alt = option_as_alt;
+    }
+
+    fn option_as_alt(&self) -> OptionAsAlt {
+        let shared_state_lock = self.shared_state.lock().unwrap();
+        shared_state_lock.option_as_alt
     }
 }