+++ src/platform_impl/orbital/window.rs
@@ -12,6 +12,7 @@ use crate::{
     error,
     platform_impl::Fullscreen,
     window,
+    window::ImePurpose,
 };
 
 use super::{
@@ -187,7 +188,7 @@ impl Window {
         //TODO: adjust for window decorations
         let (x, y): (i32, i32) = position.to_physical::<i32>(self.scale_factor()).into();
         self.window_socket
-            .write(format!("P,{},{}", x, y).as_bytes())
+            .write(format!("P,{x},{y}").as_bytes())
             .expect("failed to set position");
     }
 
@@ -206,7 +207,7 @@ impl Window {
     pub fn set_inner_size(&self, size: Size) {
         let (w, h): (u32, u32) = size.to_physical::<u32>(self.scale_factor()).into();
         self.window_socket
-            .write(format!("S,{},{}", w, h).as_bytes())
+            .write(format!("S,{w},{h}").as_bytes())
             .expect("failed to set size");
     }
 
@@ -236,7 +237,7 @@ impl Window {
     #[inline]
     pub fn set_title(&self, title: &str) {
         self.window_socket
-            .write(format!("T,{}", title).as_bytes())
+            .write(format!("T,{title}").as_bytes())
             .expect("failed to set title");
     }
 
@@ -322,6 +323,9 @@ impl Window {
 
     #[inline]
     pub fn set_ime_allowed(&self, _allowed: bool) {}
+
+    #[inline]
+    pub fn set_ime_purpose(&self, _purpose: ImePurpose) {}
 
     #[inline]
     pub fn focus_window(&self) {}